+++ iOSDeviceManager/Devices/Device.h
@@ -26,24 +26,24 @@
 + (iOSReturnStatusCode)generateXCAppDataBundleAtPath:(NSString *)path
                                            overwrite:(BOOL)overwrite;
 
-- (iOSReturnStatusCode)installApp:(Application *)app shouldUpdate:(BOOL)shouldUpdate;
+- (iOSReturnStatusCode)installApp:(Application *)app forceReinstall:(BOOL)forceReinstall;
 - (iOSReturnStatusCode)installApp:(Application *)app
                 resourcesToInject:(NSArray<NSString *> *)resourcePaths
-                     shouldUpdate:(BOOL)shouldUpdate;
+                     forceReinstall:(BOOL)forceReinstall;
 - (iOSReturnStatusCode)installApp:(Application *)app
                     mobileProfile:(MobileProfile *)profile
-                     shouldUpdate:(BOOL)shouldUpdate;
+                     forceReinstall:(BOOL)forceReinstall;
 - (iOSReturnStatusCode)installApp:(Application *)app
                     mobileProfile:(MobileProfile *)profile
                 resourcesToInject:(NSArray<NSString *> *)resourcePaths
-                     shouldUpdate:(BOOL)shouldUpdate;
+                     forceReinstall:(BOOL)forceReinstall;
 - (iOSReturnStatusCode)installApp:(Application *)app
                  codesignIdentity:(CodesignIdentity *)codesignID
-                     shouldUpdate:(BOOL)shouldUpdate;
+                     forceReinstall:(BOOL)forceReinstall;
 - (iOSReturnStatusCode)installApp:(Application *)app
                  codesignIdentity:(CodesignIdentity *)codesignID
                 resourcesToInject:(NSArray<NSString *> *)resourcePaths
-                     shouldUpdate:(BOOL)shouldUpdate;
+                     forceReinstall:(BOOL)forceReinstall;
 - (iOSReturnStatusCode)uninstallApp:(NSString *)bundleID;
 - (iOSReturnStatusCode)simulateLocationWithLat:(double)lat lng:(double)lng;
 - (iOSReturnStatusCode)stopSimulatingLocation;
@@ -54,7 +54,9 @@
 - (BOOL)launchApplicationWithConfiguration:(FBApplicationLaunchConfiguration *)configuration
                                      error:(NSError **)error;
 - (iOSReturnStatusCode)killApp:(NSString *)bundleID;
-- (BOOL)shouldUpdateApp:(Application *)app statusCode:(iOSReturnStatusCode *)sc;
+- (BOOL)shouldUpdateApp:(Application *)newApp
+           installedApp:(Application *)installedApp
+             statusCode:(iOSReturnStatusCode *)codeRef;
 - (iOSReturnStatusCode)isInstalled:(NSString *)bundleID;
 - (BOOL)isInstalled:(NSString *)bundleID withError:(NSError **)error;
 - (Application *)installedApp:(NSString *)bundleID;